Add files via upload
I made this Pomodoro timer using vanilla JavaScript. I decided I want to create a project that I would use, as I believed this would give it more personal value along with motivating me to create an end product that was useful. I used Materialize to create a user interface that was modern and simple, as I wanted the user to be able to start using the timer instantly and not get distracted from their main task, as this would contradict the overall purpose of The Pomodoro Technique. 
 Once the timer is finished it alerts the user by playing a sound and displaying a toast as visual notification. The user can then start a new timer to start a break or a new Pomodoro cycle.